+++ app/actions/todos.js
@@ -1,50 +1,43 @@
-import { createActions } from 'redux-actions';
+import { createAction } from 'redux-actions';
 import todosSource from 'sources/todos';
-import actionTypes from 'constants/todos';
 
-const {
-  LOAD_TODOS,
-  SET_TODOS,
-  ADD_TODO,
-  TOGGLE_TODO,
-  REMOVE_TODO
-} = actionTypes;
+export const LOAD_TODOS = 'LOAD_TODOS';
+export const SET_TODOS = 'SET_TODOS';
+export const ADD_TODO = 'ADD_TODO';
+export const TOGGLE_TODO = 'TOGGLE_TODO';
+export const REMOVE_TODO = 'REMOVE_TODO';
 
-const actions = createActions(
-  LOAD_TODOS,
-  SET_TODOS,
-  ADD_TODO,
-  TOGGLE_TODO,
-  REMOVE_TODO
-);
+export const loadTodos = createAction(LOAD_TODOS);
+export const setTodos = createAction(SET_TODOS);
+export const addTodo = createAction(ADD_TODO);
+export const toggleTodo = createAction(TOGGLE_TODO);
+export const removeTodo = createAction(REMOVE_TODO);
 
-const fetchTodos = () =>
+export const fetchTodos = () =>
   dispatch => {
-    dispatch(actions.loadTodos());
+    dispatch(loadTodos());
 
     return todosSource.get().then(result => {
-      dispatch(actions.setTodos(result));
+      dispatch(setTodos(result));
     });
   };
 
-const createTodo = todo =>
+export const createTodo = todo =>
   dispatch =>
     todosSource.create({ ...todo, isComplete: false }).then(result => {
-      dispatch(actions.addTodo(result));
+      dispatch(addTodo(result));
     });
 
-const updateTodo = todo =>
+export const updateTodo = todo =>
   dispatch => {
-    dispatch(actions.toggleTodo(todo));
+    dispatch(toggleTodo(todo));
 
     return todosSource.update(todo);
   };
 
-const deleteTodo = todo =>
+export const deleteTodo = todo =>
   dispatch => {
-    dispatch(actions.removeTodo(todo));
+    dispatch(removeTodo(todo));
 
     return todosSource.delete(todo);
   };
-
-export default { ...actions, fetchTodos, createTodo, updateTodo, deleteTodo };