+++ app/constants/todos.js
@@ -1,9 +1,0 @@
-import mirrorCreator from 'mirror-creator';
-
-export default mirrorCreator([
-  'LOAD_TODOS',
-  'SET_TODOS',
-  'ADD_TODO',
-  'TOGGLE_TODO',
-  'REMOVE_TODO'
-]);