+++ app/containers/application/index.jsx
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
-import sessionActions from 'actions/session';
-import modalActions from 'actions/modal';
+import { logoutUser } from 'actions/session';
+import { openModal } from 'actions/modal';
 import Application from 'components/application';
 
 const mapStateToProps = state => ({
@@ -8,9 +8,9 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = dispatch => ({
-  logout: user => dispatch(sessionActions.logoutUser(user)),
-  signin: () => dispatch(modalActions.openModal({ name: 'signin' })),
-  signup: () => dispatch(modalActions.openModal({ name: 'signup' }))
+  logout: user => dispatch(logoutUser(user)),
+  signin: () => dispatch(openModal({ name: 'signin' })),
+  signup: () => dispatch(openModal({ name: 'signup' }))
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(Application);